requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
dotenv==0.9.9
fastapi==0.115.8
h11==0.14.0
idna==3.10
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
